+++ src/pages/assets/AssetsOverviewPage.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import { graphql } from 'react-apollo'
 import WatchlistCard from '../../components/Watchlists/WatchlistCard'
-import FeaturedWatchlist from '../../components/Watchlists/FeaturedWatchlist'
 import GetWatchlists from './../../ducks/Watchlists/GetWatchlists'
 import { allSlugsForAssetTypesGQL } from './../Projects/allProjectsGQL'
 import styles from './AssetsOverview.module.scss'
@@ -48,7 +47,6 @@ const AssetsOverview = props => (
         />
       ))}
     </div>
-    <FeaturedWatchlist />
     <h4>My watchlists</h4>
     <div className={styles.flexRow}>
       <GetWatchlists
